requirements.txt
contourpy==1.2.0
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.44.0
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
